Move tests in separate file
